Refactored the code a bit.  Extended classes and moved functionality.
The ultimate goal with these changes is to make it easier to keep
changes simple.  It is already more readable.

I plan on making the find element methods a little cleaner and moving
the combination of those methods into the *Test.java version of any test
cases so pauses can be inserted easier.

Also need to find a way to check for page load that is reliable,
currently making the system waiting any less than 2 seconds causes tests
to fail due to page not being entirely loaded.